+++ jest-bundle-cjs.config.js
@@ -1,4 +1,5 @@
 module.exports = {
+    setupFilesAfterEnv: ["jest-expect-message"],
     moduleNameMapper: {
         '~\/src$': '<rootDir>/dist/cjs/index.js',
     },