+++ jest.config.js
@@ -1,4 +1,5 @@
 module.exports = {
+    setupFilesAfterEnv: ["jest-expect-message"],
     moduleNameMapper: {
         '~(.*)$': '<rootDir>/$1',
     },