+++ src/index.js
@@ -3,7 +3,7 @@ import {convert, convertFromPip, convertToPip, numToBig} from './converter.js';
 import {
     mPrefixToHex, mPrefixStrip, mToBuffer, toBuffer, addressToString, checkToString, privateToAddressString, isMinterPrefixed, isValidAddress, isValidCheck, isValidTransaction, isValidPublicKeyString,
 } from './prefix.js';
-import {getFeeValue} from './fee.js';
+import {getFeeValue, BASE_FEES} from './fee.js';
 import {
     sellCoin, sellCoinByBip, buyCoin, buyCoinByCoin,
 } from './coin-math.js';
@@ -41,6 +41,7 @@ export {
     TX_TYPE,
     txTypeList,
     normalizeTxType,
+    BASE_FEES,
     COIN_MAX_AMOUNT,
     COIN_MAX_MAX_SUPPLY,
     COIN_MIN_MAX_SUPPLY,