+++ src/variables.js
@@ -1,16 +1,15 @@
-// values in BIP
 /**
  * Maximum amount of any coin
  * @type {number}
  */
 export const COIN_MAX_AMOUNT = 10 ** 15;
 /**
- * Coin's maximum `maxSupply` value
+ * Coin's maximum `maxSupply` value (in BIP)
  * @type {number}
  */
 export const COIN_MAX_MAX_SUPPLY = COIN_MAX_AMOUNT;
 /**
- * Coin's minimum `maxSupply` value
+ * Coin's minimum `maxSupply` value (in BIP)
  * @type {number}
  */
 export const COIN_MIN_MAX_SUPPLY = 1;